- create applications , application events , contacts and contact events tables , migration and models - adjust the kpi indicator -add the applications , application events , contacts and contact events tables , migration and API's

+++ database/migrations/2018_10_03_101950_create_employees_table.php
@@ -16,7 +16,7 @@ class CreateEmployeesTable extends Migration {
 		{
 			$table->integer('id', true);
 			$table->string('name', 100);
-			$table->integer('zoho_id');
+			$table->bigInteger('zoho_id');
 			$table->integer('office_id')->index('office_id');
 			$table->timestamp('created_at')->default(DB::raw('CURRENT_TIMESTAMP'));
 		});
